requirements.txt
aiohttp>=3.9.0
asyncio-throttle>=1.0.2
beautifulsoup4>=4.12.0
click>=8.1.0
google-generativeai>=0.8.0
lxml>=4.9.0
pydantic>=2.5.0
PyYAML>=6.0
rich>=13.7.0
tenacity>=8.2.0
urllib3>=2.1.0
fake-useragent>=1.4.0
aiofiles>=23.2.0
python-dotenv>=1.0.0
